+++ src/main/java/com/ruoyi/project/vocpurchase/domain/dto/UpdateVocPurchaseRequestDto.java
@@ -5,7 +5,6 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -30,10 +29,10 @@ public class UpdateVocPurchaseRequestDto extends BaseRequestDto
     private String purchaseType;
 
     @ApiModelProperty(value = "采购总数")
-    private int totalAmount;
+    private Integer totalAmount;
 
     @ApiModelProperty(value = "采购总价")
-    private BigDecimal totalPrice;
+    private Double totalPrice;
 
     @ApiModelProperty(value = "采购商品信息列表")
     private List<UpdateVocPurchaseItemRequestDto> updateVocPurchaseItemRequestDtoList;
